Initial version of GPS SPS schema featuring DFDL parity layers.
This is not functioning. It is a "mock up" of how a DFDL schema
that verifies parity bits, and recomputes them on unparse would work.

This is intended to support review of this concept of how DFDL schemas
would capture parity calculations. The basic idea is that the
existing layer transforms functionality would be extended to be able
to both read and write DFDL variables. That in conjunction with use of
dfdl:newVariableInstance, dfdl:outputValueCalc, and dfdl:assert tests is
sufficient.

So far only the subframe type 1 is described.